requirements.txt
Django==3.2.8
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.5
qrcode==7.3.1
drf-yasg==1.20.0

